requirements.txt
Telethon==1.37.0
PyYAML==6.0.2
python-dotenv==1.0.1
